Move multidimensional code later in boot sequence
